Consolidate torchrun publication modules
